+++ warm_alarm_ios/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Add notification authorization and notification-settings remediation.
+
 # 0.1.0+1
 
 - Initial release of this plugin.